update mklinks.sh, check if ~/.config/neovim/init.vim doesn't exist before writing

+++ mklnks.sh
@@ -9,7 +9,7 @@ linkme ()
         ln -s $RC_DIR/"$1" $HOME/."$1"
         echo "Linked $1 to $HOME/.$1"
     else
-        echo "Skipped linking $1, $HOME/.$1 already exists"
+        echo "Skipped linking $1, $HOME/.$1 already exists, won't overwrite."
     fi
 }
 
@@ -26,10 +26,14 @@ linkme vim
 echo 'linking vimrc for neovim, if vimrc exists'
 if [ -e $HOME/.vim/vimrc ] ; then
 	echo 'vimrc exists'
-	ln -s $HOME/.vim $HOME/.config/nvim
-    echo "Linked $HOME/.vim to $HOME/.config/nvim"
-    ln -s $HOME/.vimrc $HOME/.config/nvim/init.vim
-    echo "Linked $HOME/.vimrc $HOME/.config/nvim/init.vim"
+    if [ ! -e $HOME/.config/nvim/init.vim ] ; then
+        ln -s $HOME/.vim $HOME/.config/nvim
+        echo "Linked $HOME/.vim to $HOME/.config/nvim"
+        ln -s $HOME/.vimrc $HOME/.config/nvim/init.vim
+        echo "Linked $HOME/.vimrc $HOME/.config/nvim/init.vim"
+    else
+        echo "$HOME/.config/nvim/init.vim already exists, won't overwrite."
+    fi
 else
     echo "Skpped linking $HOME/.config/nvim , $HOME/.vim/vimrc doesn't exist"
 fi
